// src/api/request/code.ts
/**
 * @description: 错误码
 * @param {REQUEST_ERROR} REQUEST_ERROR 请求错误 400
 * @param {TOKEN_EXPIRED} TOKEN_EXPIRED Token过期，请重新登录 401
 * @param {ACCESS_DENY} ACCESS_DENY 拒绝访问 403
 * @param {NOT_FOUND} NOT_FOUND 请求地址出错 404
 * @param {SERVER_INTER_ERROR} SERVER_INTER_ERROR 服务器内部错误 500
 * @param {SERVICE_NOT_IMPLEMENT} SERVICE_NOT_IMPLEMENT 服务未实现 501
 * @param {GATEWAY_ERROR} GATEWAY_ERROR 网关错误 502
 * @param {SERVICE_UNAVAILABLE} SERVICE_UNAVAILABLE 服务不可用 503
 * @param {GATEWAY_TIMEOUT} GATEWAY_TIMEOUT 服务不可用 504
 * @param {HTTP_NOT_SUPPORT} HTTP_NOT_SUPPORT HTTP版本不受支持 505
 * @return {*}
 */
export enum EErrorCode {
    REQUEST_ERROR = 400,
    TOKEN_EXPIRED = 401,
    ACCESS_DENY = 403,
    NOT_FOUND = 404,
    SERVER_INTER_ERROR = 500,
    SERVICE_NOT_IMPLEMENT = 501,
    GATEWAY_ERROR = 502,
    SERVICE_UNAVAILABLE = 503,
    GATEWAY_TIMEOUT = 504,
    HTTP_NOT_SUPPORT = 505,
}
  